cloudcomposerdiff/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudcomposerdiff"
version = "1.0.0"
description = "Determine the diff between two cloud composer environments"
dependencies = [
    "google-cloud-orchestration-airflow",
    "click",
    "rich",
]

[project.scripts]
cloudcomposerdiff = "cloudcomposerdiff.__main__:main"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest", "flake8"]

[tool.bumpver]
current_version = "1.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/cloudcomposerdiff/__init__.py" = ["{version}"]